Step_3/requirements.txt
flask==3.0.1
line-bot-sdk==3.7.0
python-dotenv==1.0.0
azure-data-tables==12.5.0
python-dateutil==2.8.2